simplified switch reading
I am not dealing with finicky buttons like the button_ISR_LED project
was, and I am not turning ISRs on and off in response.  All I need to do
is read the current state and turn an LED on or off for each switch, so
all that stuff can happen in main.

Also renamed the "ISR_handlers" source file in the button_ISR_LED
project to "my_ISR_handlers".