building a framework for common functionality
So far, I have my delay timer for delays in non-RTOS functions, a
makeshift runtime alarm that blinks the onboard LEDs of my MX4cK in
rapid and alternating patterns, and the beginnings of an I2C handler.

The I2C handler's organization is more difficult than I initially
intended.  I need to think more on how this should be organized to
account for a single I2C handle, but multiple devices potentially being
on the same I2C bus.