making progress on the I2C handler
I have modified how the I2C handler is called and what it's
responsibilities will be.  I decided that, for simplicity, it is better
to integrate new I2C functionality, such as communicating with a
particular pmod, into the I2C handler instead of making it a separate
handler.  The initializations got too messy when I tried to figure out
how to make sure that initializations happened in the right order.