modified the I2C handler to handle timeouts
Rather than hanging forever in while(...) loops, as it did previously,
the I2C handler now makes use of the delay timer to figure out how long
it has been hanging and to jump out of loops if it has stalled for too
long.

Also modified my delay timer to be a class.