making some progress on the I2C line
I made an error when I tried to combine "start transfer with restart"
with "start transfer without restart".  The latter waited for the I2C
line to become idle, anticipating that the transfer had already been
stopped by whatever code had previously used it.  The former code does
not anticipate this, but in my code I had "start transfer with restart"
waiting on the bus that would never become idle because nothing had sent
a stop signal.

Now I am deciphering why the bus is not stopping after an attempted
reading of the accelerometer.