began working on internal servo control
I started working on a basic servo control project.  I am using a
decade-old or more servo that came with an unused RC plane that my
grandpa bought, and by 3.3 IO pins are able to drive it.

I also discovered that that the 5V output of the RC kit's ESC BIC worked
like a charm in either powering the whole MX4cK or power the servo
lines.  I haven't tried doing it with both because the BIC is only rated
for 2A.

This program uses a custom 0.1 millisecond timer to vary the pulse width
between the standard 1000 microsecond and 2000 microsecond range for
servos.  On my servo, this only gives me a 90 degree range instead of
the full 180, but 90 is good, and customizing the pulse width would
require me to go outside the standards, which may not be okay.