timer stuff abandoned, continuing wifi
The timer shenanigans were bizarre, so I eventually abandoned them and
moved on to the wifi communication.  Due to the block nature of the
TCPIP stack servicing when it first connects to the network, the stack
services must be in either an RTOS task or in main()'s while(1) loop.

I chose the latter, and am now attempting to migrate some code from
FreeRTOS in a friend's project to non-RTOS code.  Doing this will
simplify the presentation and make the underlying operations easier to
understand.