got sockets working, now communication
Tera Term doesn't complain about logging in anymore, but the thing is
getting stuck when I try to send and receive information.  In
particular, it gets stuck way down in the TCPIP stack when it waits for
an interrupt from the wifi pmod's SPI pin.

I got a tip from someone on Reddit: Don't call functions from within
interrupts because weird stuff can happen.

I am calling my "TCPIP send and receive" function from within my
interrupt, and perhaps because of the way this PIC32 processor works and
the way that the SPI interrupt is configured, the SPI interrupt can't
fire while my timer interrupt (the one that calls the "TCPIP send and
receive" function) is active.

Committing before resorting to a "function queue" design.