yay! data is being sent and received!
Hooray!  The interrupt was the problem.  Now, instead of calling the
"TCPIP send and receive" function directly, my timer interrupt calls a
function and adds the function pointer to a queue that is subsequently
called by the while(1) loop in main().

Now I will begin taking all of this mess and putting it into a simple
demo project, where it will be cleaned up and prepared to be used by
other people.