made some adjustments to send-receive
I had to adjust the way that the received data was spit back out so that
it would differ from the transmitted message.  As it was, in Tera Term,
hitting the return key when the "local echo" was turned on entered the
"return carriage" character '\r' without a newline '\n' character, so
the reprinted and received text simply overwrote what was sent to the
microcontroller.

I added buffers for sending and receiving in addition to the CLS message
buffer.  Now the returned message says: "--You said: 'whatever message
you sent'".