tearing out MHCP_TCPIP.h
I do not like the idea of a lot of the details of TCPIP and wifi setup
taking place in a header file that is dominated by macros, so I am
moving the functionality of MHCP_TCPIP.h into my TCPIP framework.

I am also tearing apart that functionality and removing support for
anything that is not my wifi module.  I do not need those additional
checks in my code because there is no alternative hardware that I need
to support.