trimmed up the readme
I made a few lines of code from main() to the TCPIP and wifi init
function, then commented it out because it didn't seem to affect code
operation.

I also deleted the startup delay, which didn't seem to do anything.

I added some notes at the end of the readme about where to move code to
if you want to use an RTOS.